题目列表功能-后端开发02

+++ oj-common/oj-common-core/src/main/java/com/bite/common/core/domain/TableDataInfo.java
@@ -48,7 +48,7 @@ public class TableDataInfo {
         return rspData;
     }
 
-    public static TableDataInfo success(List<?> list, int total) {
+    public static TableDataInfo success(List<?> list, long total) {
         TableDataInfo rspData = new TableDataInfo();
         rspData.setCode(ResultCode.SUCCESS.getCode());
         rspData.setRows(list);
